Feature/#727

+++ src/TD.Web/TD.Web.Public/TD.Web.Public.Contrats/Requests/Cart/CheckoutRequestBase.cs
@@ -6,7 +6,7 @@
         public string? Name { get; set; }
         public string? Mobile { get; set; }
         public string? Note { get; set; }
-        public int StoreId { get; set; }
+        public short StoreId { get; set; }
         public int PaymentTypeId { get; set; }
     }
 }
